requirements.txt
docling>=2.0.0
pypdf>=4.0.0
pdfplumber>=0.11.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
umap-learn>=0.5.5
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
nltk>=3.8.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.66.0
pyyaml>=6.0
